Auto-replace ASCII metric characters with `_` when needed
Prometheus metric names must match the regex `[a-zA-Z_:][a-zA-Z0-9_:]*`.
Today, we only automatically replace `/` with `_`, and any other
non-compliant character will trigger a panic in the Prometheus
libraries.

Some characters like `-` are also common in metric names, and this
limitation adds unnecessary configuration overhead.

This patch makes the code auto-replace all ASCII characters that aren't
compatible with Prometheus' limits with `_`.

Note that non-ASCII characters will still trigger a panic. There is no
good way to handle all of them automatically, as they can't be all reasonably
mapped to ascii. In the future, we may handle _some_ of them automatically when
possible. But for now, forcing the users to be explicit is the safest option,
and also ensures forwards compatibility.

Thanks to mschirrmeister@github for reporting this issue in
https://github.com/albertito/prometheus-expvar-exporter/issues/1.